clinic_rest/requirements.txt
Django
djangorestframework
jwt
redis
pyjwt
django-cors-headers